refactor: dashboard 페이지에서 favorite 페칭 => indicators component 에서 fetching
