+++ src/pages/dashboard/index.tsx
@@ -1,65 +1,16 @@
 import clsx from 'clsx';
 import { useState } from 'react';
-import queryKey from '@/const/queryKey';
 import { roboto, poppins } from '../_app';
 import Menu from '@/components/menu/Menu';
-import app from '@/firebase/firebaseConfig';
 import styles from './Dashboard.module.scss';
-import { useQuery } from '@tanstack/react-query';
 import MyContext from '@/components/myContext/MyContext';
-import { getDatabase, ref, get } from 'firebase/database';
 import Dashheader from '@/components/dashheader/DashHeader';
 import Indicators from '@/components/indicators/Indicators';
-
-type Favorite = {
-	title: string;
-	seriesId: string;
-	categoryId: number;
-};
-
-const fetchFavoritesByUserId = async (userId: number): Promise<Favorite[]> => {
-	const db = getDatabase(app);
-	const favoritesRef = ref(db, `/user/favorite/${userId}`);
-	const favoriteArray: Favorite[] = [];
-
-	await get(favoritesRef)
-		.then(snapshot => {
-			if (snapshot.exists()) {
-				const favoritesData = snapshot.val();
-				favoriteArray.push(favoritesData);
-			} else {
-				console.log('No data available');
-			}
-		})
-		.catch(error => {
-			console.error(error);
-		});
-
-	return favoriteArray;
-};
 
 export default function Dashboard() {
 	const [Tabs] = useState(['Indicators', 'MyContext']);
 	const [TabsIndex, setSelectedIdx] = useState(0);
-
 	const [CategoryIndex, setCategoryIndex] = useState(0);
-	const { data: favorites, isSuccess } = useQuery({
-		queryKey: [queryKey.favorite, 1],
-		queryFn: async () => {
-			const db = getDatabase(app);
-			const favoriteDocRef = ref(db, `/user/favorite/${1}`);
-			const snapshot = await get(favoriteDocRef);
-			const favoriteArray: Favorite[] = [];
-
-			if (snapshot.exists()) {
-				snapshot.forEach(childSnapshot => {
-					favoriteArray.push(childSnapshot.val());
-				});
-			}
-
-			return favoriteArray;
-		}
-	});
 
 	return (
 		<>
@@ -70,8 +21,8 @@ export default function Dashboard() {
 				<section>
 					<Dashheader Tabs={Tabs} TabsIndex={TabsIndex} />
 
-					{Tabs[TabsIndex] === 'Indicators' && isSuccess ? (
-						<Indicators favorites={favorites} CategoryIndex={CategoryIndex} setCategoryIndex={setCategoryIndex} />
+					{Tabs[TabsIndex] === 'Indicators' ? (
+						<Indicators CategoryIndex={CategoryIndex} setCategoryIndex={setCategoryIndex} />
 					) : (
 						<MyContext />
 					)}
